+++ .env.example
@@ -2,4 +2,3 @@ VITE_NODE_ENV=              # development or production or test
 VITE_API_ROOT_DEV=          # Root API for development
 VITE_API_ROOT_PROD=         # Root API for production
 VITE_API_ROOT_LOCAL=        # Root API for local
-BACKEND_URL_LOCAL  =  3000          # Backend URL
