+++ src/config/env.ts
@@ -2,4 +2,3 @@ export const API_ROOT_DEV = import.meta.env.VITE_API_ROOT_DEV;
 export const API_ROOT_PROD = import.meta.env.VITE_API_ROOT_PROD;
 export const API_ROOT_LOCAL = import.meta.env.VITE_API_ROOT_LOCAL;
 export const NODE_ENV = import.meta.env.VITE_NODE_ENV;
-export const BACKEND_URL = import.meta.env.BACKEND_URL_LOCAL;
